+++ src/Entity/EntityManager.hpp
@@ -15,17 +15,9 @@ class EntityManager {
 
     BE_NON_COPY_CONSTRUTIBLE(EntityManager)
 
-    void Update(f64 delta);
+    void Update(f32 delta);
     Entity CreateEntity(const std::string &name);
     void DestroyEntity(Entity entity);
-
-    // template <typename T, typename... Args>
-    // Entity CreateScriptableEntity(const std::string &name, Args &&...args) {
-    //     auto e = CreateEntity(name);
-    //     e.AddComponent<Components::Script>().Bind<T, Args...>(
-    //         e, std::forward<Args>(args)...);
-    //     return e;
-    // }
 
    private:
     entt::registry m_Registry;
