+++ src/Entity/ScriptableEntity.hpp
@@ -15,7 +15,7 @@ class ScriptableEntity {
 
    protected:
     virtual void OnCreate() = 0;
-    virtual void OnUpdate(f64) = 0;
+    virtual void OnUpdate(f32) = 0;
     virtual void OnDestroy() = 0;
 
     void Destroy();
